Add WIP Message for Unlinked Options
As this covid-19 dungeon is work in progress, there are many options which are not linked to any files. So, I have added a `WIP.md` file and linked all unlinked files to this new file.

If you are interested in completing this dungeon, please see [Contribution Guidelines](../../../README.md#contribution-guidelines)