Add Decision Tree Reference in README (#136)
* Fix All Broken Links in Covid-19 Dungeon

There were a few broken links in the covid-19 dungeon. Also, no documents were linked to the `3/3.md`. So, added a link for that in `1/1-CB.md`.

* Add Reference to Decision Tree in README

Currently, there is no way of knowing that a decision tree exists in this repo. This commit adds a reference to it in `README.md > Contribution Guidelines > Dungeon Structure`.