Fix Dungeon Title in Start Game File  (#137)
* Fix All Broken Links in Covid-19 Dungeon

There were a few broken links in the covid-19 dungeon. Also, no documents were linked to the `3/3.md`. So, added a link for that in `1/1-CB.md`.

* Fix Dungeon Title in Start Game File

Currently, the `start-new-game.md` file references the `covid-19` dungeon as `it's 2020`. This doesn't make any sense as the pandemic is not over in 2020. So, changed it to `Save yourself from the viruses!`